+++ tests/testthat/test-createDSR.R
@@ -53,28 +53,28 @@ sapply(studies, function(study) {
     
   })
   
-  test_that("overwrite a DSR ", {
-    dsrData<-data.frame(area = c("a", "b"), unit = c(52,36), nominalCapacity = c(956, 478), marginalCost = c(52, 65), hour = c(3, 7))
+  # test_that("overwrite a DSR ", {
+    # dsrData<-data.frame(area = c("a", "b"), unit = c(52,36), nominalCapacity = c(956, 478), marginalCost = c(52, 65), hour = c(3, 7))
     
-    expect_error(suppressWarnings(createDSR(dsrData)), "The link a - a_dsr_3h already exist, use overwrite.")
+    # expect_error(suppressWarnings(createDSR(dsrData)), "The link a - a_dsr_3h already exist, use overwrite.")
     
-    createDSR(dsrData, overwrite = TRUE)
-    linkADsr <- "a - a_dsr_3h"
-    linkBDsr <- "b - b_dsr_7h"
-    expect_true(linkADsr %in% getLinks())
-    expect_true(linkBDsr %in% getLinks())
-    capaLink<-antaresRead::readInputTS(linkCapacity = c("a - a_dsr_3h", "b - b_dsr_7h"), showProgress = FALSE)
-    expect_equal(unique(capaLink[link==linkADsr, transCapacityIndirect]), dsrData[dsrData$area=="a",]$unit*dsrData[dsrData$area=="a",]$nominalCapacity)
-    expect_equal(unique(capaLink[link==linkBDsr, transCapacityIndirect]), dsrData[dsrData$area=="b",]$unit*dsrData[dsrData$area=="b",]$nominalCapacity)
+    # createDSR(dsrData, overwrite = TRUE)
+    # linkADsr <- "a - a_dsr_3h"
+    # linkBDsr <- "b - b_dsr_7h"
+    # expect_true(linkADsr %in% getLinks())
+    # expect_true(linkBDsr %in% getLinks())
+    # capaLink<-antaresRead::readInputTS(linkCapacity = c("a - a_dsr_3h", "b - b_dsr_7h"), showProgress = FALSE)
+    # expect_equal(unique(capaLink[link==linkADsr, transCapacityIndirect]), dsrData[dsrData$area=="a",]$unit*dsrData[dsrData$area=="a",]$nominalCapacity)
+    # expect_equal(unique(capaLink[link==linkBDsr, transCapacityIndirect]), dsrData[dsrData$area=="b",]$unit*dsrData[dsrData$area=="b",]$nominalCapacity)
     
-    #edit spinning
-    optsRes <- createDSR(dsrData, overwrite = TRUE, spinning = 3)
-    clusterList <- antaresRead::readClusterDesc(opts = optsRes)
-    expect_equal(as.character(clusterList[area == "a_dsr_3h"]$cluster), "a_dsr_3h_cluster")
-    expect_equal(as.character(clusterList[area == "a_dsr_3h"]$group), "Other")
-    expect_equal(as.double(clusterList[area == "a_dsr_3h"]$spinning), 3)
+    # #edit spinning
+    # optsRes <- createDSR(dsrData, overwrite = TRUE, spinning = 3)
+    # clusterList <- antaresRead::readClusterDesc(opts = optsRes)
+    # expect_equal(as.character(clusterList[area == "a_dsr_3h"]$cluster), "a_dsr_3h_cluster")
+    # expect_equal(as.character(clusterList[area == "a_dsr_3h"]$group), "Other")
+    # expect_equal(as.double(clusterList[area == "a_dsr_3h"]$spinning), 3)
     
-  })
+  # })
   
   test_that("test input data DSR", {
     #area
@@ -104,33 +104,33 @@ sapply(studies, function(study) {
     expect_error(createDSR(dsrData, overwrite = TRUE, spinning = NULL), "spinning is set to NULL")
   })
   
-  test_that("getCapacityDSR and editDSR", {
-    dsrData<-data.frame(area = c("a", "b"), unit = c(50,40), nominalCapacity = c(200, 600), marginalCost = c(52, 65), hour = c(3, 7))
-    createDSR(dsrData, overwrite = TRUE)
+  # test_that("getCapacityDSR and editDSR", {
+    # dsrData<-data.frame(area = c("a", "b"), unit = c(50,40), nominalCapacity = c(200, 600), marginalCost = c(52, 65), hour = c(3, 7))
+    # createDSR(dsrData, overwrite = TRUE)
     
-    expect_equal(getCapacityDSR("a"),  dsrData[dsrData$area=="a",]$nominalCapacity *  dsrData[dsrData$area=="a",]$unit )
-    expect_equal(getCapacityDSR("b"),  dsrData[dsrData$area=="b",]$nominalCapacity *  dsrData[dsrData$area=="b",]$unit )
+    # expect_equal(getCapacityDSR("a"),  dsrData[dsrData$area=="a",]$nominalCapacity *  dsrData[dsrData$area=="a",]$unit )
+    # expect_equal(getCapacityDSR("b"),  dsrData[dsrData$area=="b",]$nominalCapacity *  dsrData[dsrData$area=="b",]$unit )
     
-    optsRes<-editDSR(area = "a", 
-                     unit = 2, 
-                     nominalCapacity = 500,
-                     marginalCost = 40,
-                     spinning = 50)
+    # optsRes<-editDSR(area = "a", 
+                     # unit = 2, 
+                     # nominalCapacity = 500,
+                     # marginalCost = 40,
+                     # spinning = 50)
     
-    #change for "a" but not for "b" 
-    expect_equal(getCapacityDSR("a"), 2 * 500)
-    expect_equal(getCapacityDSR("b"),  dsrData[dsrData$area=="b",]$nominalCapacity *  dsrData[dsrData$area=="b",]$unit )
-    #get the new values
-    clusterList <- antaresRead::readClusterDesc(opts = optsRes)
-    dsrName <- "a_dsr_3h"
-    expect_equal(as.character(clusterList[area == dsrName]$cluster), paste0(dsrName, "_cluster"))
-    expect_equal(as.character(clusterList[area == dsrName]$group), "Other")
-    expect_equal(clusterList[area == dsrName]$enabled, TRUE)
-    expect_equal(clusterList[area == dsrName]$unitcount, 2)
-    expect_equal(clusterList[area == dsrName]$spinning, 50)
-    expect_equal(clusterList[area == dsrName]$nominalcapacity, 500)
-    expect_equal(clusterList[area == dsrName]$marginal.cost, 40)
-  })
+    # #change for "a" but not for "b" 
+    # expect_equal(getCapacityDSR("a"), 2 * 500)
+    # expect_equal(getCapacityDSR("b"),  dsrData[dsrData$area=="b",]$nominalCapacity *  dsrData[dsrData$area=="b",]$unit )
+    # #get the new values
+    # clusterList <- antaresRead::readClusterDesc(opts = optsRes)
+    # dsrName <- "a_dsr_3h"
+    # expect_equal(as.character(clusterList[area == dsrName]$cluster), paste0(dsrName, "_cluster"))
+    # expect_equal(as.character(clusterList[area == dsrName]$group), "Other")
+    # expect_equal(clusterList[area == dsrName]$enabled, TRUE)
+    # expect_equal(clusterList[area == dsrName]$unitcount, 2)
+    # expect_equal(clusterList[area == dsrName]$spinning, 50)
+    # expect_equal(clusterList[area == dsrName]$nominalcapacity, 500)
+    # expect_equal(clusterList[area == dsrName]$marginal.cost, 40)
+  # })