+++ tests/testthat/test-editClusterST.R
@@ -46,18 +46,27 @@ test_that("edit st-storage clusters (only for study >= v8.6.0" , {
                                        group = "Other1",
                                        add_prefix = FALSE, 
                                        opts = opts_test), 
-                         regexp = "'casper' doesn't exist,")
+                         regexp = "'casper' does not exist")
   
   ## default edition cluster ----
     # if all parameters are NULL => no edition of ini and data .txt
   testthat::expect_warning(editClusterST(area = area_test, 
                                          cluster_name = levels(st_clusters$cluster)[1],
+                                         add_prefix = FALSE,
                                          opts = opts_test), 
                            regexp = "No edition for 'list.ini' file")
   
   ## edit list ini ----
     # edit only group value
   name_cluster_test <- levels(st_clusters$cluster)[1]
+  # case insensitive
+  expect_no_error(editClusterST(area = toupper(area_test),
+                                cluster_name = toupper(name_cluster_test),
+                                group = "Other5",
+                                add_prefix = FALSE,
+                                storage_parameters = list("efficiency" = 0.789),
+                                opts = opts_test))
+
   opts_test <- editClusterST(area = area_test, 
                              cluster_name = name_cluster_test,
                              group = "Other2",